+++ core/shared/src/main/scala-2/me/shadaj/scalapy/py/Dynamic.scala
@@ -1,0 +1,7 @@
+package me.shadaj.scalapy.py
+
+import scala.language.dynamics
+
+@native trait Dynamic extends Any with AnyDynamics
+
+object Dynamic extends DynamicGlobal