+++ core/shared/src/main/scala-3/me/shadaj/scalapy/py/Object.scala
@@ -1,0 +1,3 @@
+package me.shadaj.scalapy.py
+
+@native class Object extends Any