+++ core/shared/src/main/scala/me/shadaj/scalapy/py/Any.scala
@@ -2,14 +2,14 @@ package me.shadaj.scalapy.py
 
 import scala.language.dynamics
 import scala.collection.mutable
+import scala.language.implicitConversions
 
 import me.shadaj.scalapy.interpreter.{CPythonInterpreter, PyValue}
 import me.shadaj.scalapy.readwrite.{Reader, Writer}
 
-trait Any { self =>
+trait Any extends AnyRawValue { self =>
   private var cleaned = false
   
-  private[scalapy] val rawValue: PyValue
   private[scalapy] def value: PyValue = {
     if (cleaned) {
       throw new IllegalAccessException("The Python value you are try to access has already been released by a call to py.Any.del()")
@@ -34,13 +34,7 @@ trait Any { self =>
   override def hashCode(): Int = value.hashCode()
 }
 
-object Any {
-  def populateWith(v: PyValue): Any = {
-    new Any {
-      val rawValue = v
-    }
-  }
-
+object Any extends AnyPopulateWith {
   implicit def from[T](v: T)(implicit writer: Writer[T]): Any = {
     Any.populateWith(writer.write(v))
   }