+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## vNEXT
 ### Highlights :tada:
++ Add initial support for Scala 3. To create static facades use `class` instead of `trait` in Scala 3. Also, `@PyBracketAccess` currently is not supported in this version.
 + Add support for bracket syntax to facades. The annotation `@PyBracketAccess` can be used on methods to mark them as representing bracket access on an object. The target method must have one (to read the value) or two parameters (to update the value) ([PR #194](https://github.com/shadaj/scalapy/pull/194)).
 + The version of the Python native library can now be controlled with the `scalapy.python.library` system property ([PR #198](https://github.com/shadaj/scalapy/pull/198))
 + Enable running `Py_SetProgramName` with user provided input prior to `Py_Initialize` to set the correct paths to Python run-time libraries. Input to `Py_SetProgramName`, the Python interpreter executable, can be controlled with either the `scalapy.python.programname` system property or the `SCALAPY_PYTHON_PROGRAMNAME` environment variable ([PR #200](https://github.com/shadaj/scalapy/pull/200))